refactor(auth): enclose safe-redirect in a module directory with a barrel
Group the middleware, its pure resolver, and their specs under
middlewares/safe-redirect/ with an index.ts barrel that re-exports only the public
API (registerSafeRedirectFactory + ResWithSafeRedirect). The pure resolver
(resolveSafeRedirect) and its types stay internal to the module.

Also switch the factory from a default to a named export (resolves the
noDefaultExport lint) and update the single consumer in express-init.js.

No behaviour change; 20 tests still green.

Refs #11248
