+++ apps/app/src/server/middlewares/safe-redirect/index.ts
@@ -1,0 +1,6 @@
+// Public API of the safe-redirect middleware module.
+// Internals (the pure resolveSafeRedirect() and its types in ./target) stay unexported.
+export {
+  type ResWithSafeRedirect,
+  registerSafeRedirectFactory,
+} from './middleware';
